+++ schema.dart
@@ -321,6 +321,10 @@ class AndroidTakeCameraPhotoRequest {
   late int? timeoutMillis;
 }
 
+class AndroidTakeNativeScreenshotRequest {
+  late String tag;
+}
+
 class IOSTakeCameraPhotoRequest {
   IOSSelector? shutterButtonSelector;
   IOSSelector? doneButtonSelector;
@@ -423,6 +427,9 @@ abstract class AndroidAutomator<AndroidServer, DartClient> {
   void enterText(AndroidEnterTextRequest request);
   void waitUntilVisible(AndroidWaitUntilVisibleRequest request);
   void swipe(AndroidSwipeRequest request);
+
+  // screenshots
+  void takeNativeScreenshot(AndroidTakeNativeScreenshotRequest request);
 
   // services
   void enableLocation();